+++ lib/KFCWebFramework/src/FormUIRenderHtml.cpp
@@ -156,15 +156,18 @@ namespace FormUI {
         } break;
 
         default: {
-            // append colon if label is not marked as raw output
-            if (_label.charAt(0) != 0xff && !String_endsWith(_label, ':')) {
-                _label += ':';
-            }
-            output.printf_P(PSTR("<div class=\"form-group\"><label for=\"%s\">%s</label>" FORMUI_CRLF),
-                name,
-                _encodeHtmlEntities(_label, false, output)
-            );
+            output.printf_P(PSTR("<div class=\"form-group\">" FORMUI_CRLF));
+            if (_label.length()) {
+                // append colon if label is not marked as raw output
+                if (_label.charAt(0) != 0xff && !String_endsWith(_label, ':')) {
+                    _label += ':';
+                }
+                output.printf_P(PSTR("<label for=\"%s\">%s</label>" FORMUI_CRLF),
+                    name,
+                    _encodeHtmlEntities(_label, false, output)
+                );
 
+            }
             if (_suffix.length()) {
                 output.printf_P(PSTR("<div class=\"input-group\">" FORMUI_CRLF));
             }